+++ task/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Tap non-leaf project filters to fold/hide subproject filters
   + Non-leaf projects indicated by contributing task count in parentheses
   + Unfortunately, toggling a non-leaf project filter is less convenient
+- Breaking change: tag editing UI buttons limited to tags from pending tasks
 
 # 0.2.5